token boxes added to token modal
